Added PasskeysEligibility util class
